adding form to delete and create product and adding navbar to the project,fixing delete funcion on backend to not have a constraint key error

+++ backend/src/models/productModel.js
@@ -13,7 +13,8 @@ class ProductModel {
     if (error) throw error;
 
     const product_id = data.id;
-    await StockBalanceModel.create({ product_id, quantity:0 })
+    const quantity = data.quantity || 0;
+    await StockBalanceModel.create({ product_id, quantity });
 
     return data;
   }
@@ -25,8 +26,11 @@ class ProductModel {
   }
 
   static async delete(id) {
+    const {error: balanceError} = await supabase.from('stock_balance').delete().eq('product_id', id);
+    if (balanceError) throw balanceError;
     const { error } = await supabase.from('products').delete().eq('id', id);
     if (error) throw error;
+    
   }
 }
